% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{journeys_clean}
\alias{journeys_clean}
\title{Example Oyster Journey History File (Cleaned)}
\format{An object of class \code{"data.frame"}.}
\source{
Data generated and received originally from (\href{https://tfl.gov.uk/}{Transport for London}).
}
\usage{
data(journeys_clean)
}
\description{
An anonymised example of a single month's journey history data received
originally as a CSV from Transport for London, read with \code{"oystr::oy_read()"}
and cleaned with \code{"oystr::oy_read()"}.
}
\examples{
data(journeys_clean)
}
\keyword{datasets}
