% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{journeys_read}
\alias{journeys_read}
\title{Example Oyster Journey History File (Read In)}
\format{An object of class \code{"data.frame"}.}
\source{
Data generated and received originally from (\href{https://tfl.gov.uk/}{Transport for London}).
}
\usage{
data(journeys_read)
}
\description{
An anonymised example of a single month's journey history data received
originally as a CSV from Transport for London and read with \code{"oystr::oy_read()"}.
}
\examples{
data(journeys_read)
}
\keyword{datasets}
