% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle.R
\name{oy_clean}
\alias{oy_clean}
\title{Clean Oyster journey history data}
\usage{
oy_clean(x)
}
\arguments{
\item{x}{A data.frame object containing Oyster history journey.}
}
\value{
A data.frame object.
}
\description{
Process a data.frame object containing Oyster journey history
    data. Could be a raw file as received by email from Transport for London
    or the output from oy_read().
}
