% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oy_history}
\alias{oy_history}
\title{Example Oyster Journey History File}
\format{An object of class \code{"dataframe"}.}
\source{
\href{https://tfl.gov.uk/}{Transport for London}.
}
\usage{
data(oy_history)
}
\description{
An example of a single month's journey history data received originally as
a CSV from Transport for London. Location names replaced.
}
\examples{
data(oy_history)
}
\references{
Generated by and received from (\href{https://tfl.gov.uk/}{Transport for London}).
}
\keyword{datasets}
