% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{oy_plot}
\alias{oy_plot}
\title{Simple Time Series Plots of Train Journey Data}
\usage{
oy_plot(
  data,
  x_var = "datetime_start",
  y_var = c("journey_duration", "credit", "balance"),
  weekdays = FALSE
)
}
\arguments{
\item{data}{Data frame output from \code{oy_clean()}.}

\item{x_var}{The name of the variable from \code{data} that you want on the x axis}

\item{y_var}{The name of the continuous variable frm \code{data} for the y axis.}

\item{weekdays}{Restrict the output to Monday to Friday.}
}
\value{
A data.frame object.
}
\description{
Plot train journey information (duration, charge, credit and
     balance) over time.
}
