% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{oy_read}
\alias{oy_read}
\title{Read Oyster-history CSV files from a folder}
\usage{
oy_read(path)
}
\arguments{
\item{path}{A string describing a filepath to a folder of CSVs}
}
\value{
A dataframe object.
}
\description{
Read Oyster-history CSV files from a folder
}
