% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{oy_summary}
\alias{oy_summary}
\title{Simple Summary Statistics of Travel}
\usage{
oy_summary(data, mode = "Train")
}
\arguments{
\item{data}{Data frame output from \code{oy_clean()}.}

\item{mode}{Currently restricted to "Train".}
}
\value{
A list object where each element is some kind of data summary.
}
\description{
Provide simple summaries of your Oyster journey history data.
}
