% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{oy_lineplot}
\alias{oy_lineplot}
\title{Line Plots of Oyster Journey Data}
\usage{
oy_lineplot(
  data,
  x_var = "datetime_start",
  y_var = "journey_duration",
  weekdays = FALSE,
  mode = "Train"
)
}
\arguments{
\item{data}{Data frame output from \code{oy_clean()}.}

\item{x_var}{The name of the variable from \code{data} that you want on the x axis}

\item{y_var}{The name of the continuous variable frm \code{data} for the y
axis. Choose \code{"journey_duration"} or \code{"balance"}.}

\item{weekdays}{Logical. Restrict the output to Monday to Friday with
\code{True}}

\item{mode}{Currently restricted to "Train".}
}
\value{
A data.frame object.
}
\description{
Outputs a simple time series line plots of Oyster journey data.
}
