% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{oy_read}
\alias{oy_read}
\title{Read Oyster journey history files}
\usage{
oy_read(path)
}
\arguments{
\item{path}{A string describing a filepath to a folder of CSVs}
}
\value{
A data.frame object.
}
\description{
Read one or more Oyster journey history files from a single folder
    and combine them. Assumes journey histories are raw CSV files as received
    by email from Transport for London. Ignores files that are not identified
    as Oyster history journey files.
}
